requirements.txt
supabase==1.0.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pdfminer.six==20221105
nltk==3.8.1
python-dotenv==1.0.0
spacy==3.7.2
tqdm==4.66.1
tenacity==8.2.3